[Issue 12] Adding OSGi support to the demo.
The OSGi MiceClient component supports different OSGi implementation. This release just contains the Activation and two different listeners...

This release also tres to solve the Window Adapter for the event of closing the windown and stopping the bundle before DSPComponent integration, potentially to be in the next revision. Steps to verify as follows:

- Open knopflerfish's UI - java -jar framework.jar 
- Run the build.xml to generate the correct version under the build directory;
- Choose the service, and choose PPM-8.
- Close the window and see there's a call to terminate the system.

Also adding the build.xml and the manifest for the bundle.


git-svn-id: svn://localhost/netbeams/trunk@38 559edde2-a4d4-44e0-96fc-4035e40d7396