Issue 12, Issue 15
This revision updates the build.xml script for the
MouseActions example. It also includes an image 
from the XML Schema for the Wiki documentation.

The updates to the XML schema removes the text
value of the tags, since the attributes of the
mouse info explains everything by themselves.

Finally, a new version of the Jar is included.

git-svn-id: svn://localhost/netbeams/trunk@190 559edde2-a4d4-44e0-96fc-4035e40d7396