Removing DSPWebLogger component.
There's a consistent runtime error that occurs when it tries to call DSPContextImpl.getResource() when trying to register a servlet with the OSGi framework.  The resource can't allocated and returns a null.

This occurs in both the text-console and the GUI Desktop. 

Puzzling because DSPWireTransportServer make a similar call and succeeds.

So DSPWebLogger can't be used to verify data from test components at the moment.

git-svn-id: svn://localhost/netbeams/trunk@275 559edde2-a4d4-44e0-96fc-4035e40d7396