Reorganized the jars. Moved the original Knopflerfish comm-linux_all-2.0.0.jar and serialportdevice_all-2.0.0.jar to the 'original' subdirectory.
Submitted new modified jars: comm-linux_all_ARM-2.0.0.jar and serialportdevice_all-2.0.0.jar and placed them in the 'gumstix' directory.  

Comm-linux_all_ARM-2.0.0.jar is the original jar downloaded from Knopflerfish, but with the native i386 librxtxSerial.so and librxtxParallel.so libraries replaced with the ARM ones. 

Serialportdevice_all-2.0.0.jar (same name as original) includes the javax.comm package because previous tests resulted in the package not being found when this service bundle was started.

Both need to be installed and started before installing/starting other bundles that access the gumstix's serial port.

git-svn-id: svn://localhost/netbeams/trunk@355 559edde2-a4d4-44e0-96fc-4035e40d7396