Adding the updated draft of the Data Persistence bundle. In order to test it, just add it to the config.xml and add a macher rule. Any type of message is added to the persistent layer. Finally, it includes a configuration message that determines the frequency of flushing the messages into the persistence storage.
The basic infrastructure is ready for the persistence layer.
A few problems with JPA/JDBC with OSGi: classloader problems,
Kenopflerfish problem for loading JPA, even with the hacked
version from DynamicJava.org. Maybe it's time to evaluate 
another OSGi platform such as SpringSource or Apache Felix.

Each message is identified by a SensorLocation instance, 
which contains the IP address, the longitude and latitude
of the sensor. Then, each message content is transformed
into a JSON representation of its contents (the POJO
properties). The pair is then aligned with the PropertyKey
while the DSP message is wrapped up in a PersistenMessageUnit.
Depending on the underlying Database system chosen, different
approaches my be taken to persist the objects, which will be
mapped. The current revision is going towards the direction of
a hybrid approach of relation and KVP.

http://bret.appspot.com/entry/how-friendfeed-uses-mysql
http://geekswithblogs.net/darrengosbell/articles/KVPsInDatabaseDesign.aspx

Here's the TODO list:

* Adding the persistence storage configuration as a 
bootstrap message other than the persistence.xml
file with static configuration deployed into the jar;
(Not available in this revision). This persistence
storage information follows the JPA (Java Persistence API)
and specifies the parameters for the database layer. Some
hints of how to have Hibernate as the persistence manager
was successfully described at 
http://sourceforge.net/forum/message.php?msg_id=2711171

* Filter messages by type (only persist MeasurementMessage,
for example). This could extend the persistence bundle to
be also a logger to file system, for example, capture 
trend of messages, etc.

* The persistence layer is still under revision. Evaluation
of different types of database systems (XML, KVP or Relational)
are on the go.

git-svn-id: svn://localhost/netbeams/trunk@544 559edde2-a4d4-44e0-96fc-4035e40d7396