Changed the data output from the simulator so that the Temp, SpCond, Cond, Depth, pH, etc. vary in their values by either one, two, or three decimal places.
The reason for the different decimal place changes is that given the original parameter string (which is a copy of data from the real YSI Sonde):

 21.20    193    179 5588.40   0.09   0.084   0.059  7.98   -79.6   99.5   8.83     0.4     8.7
 
- parameters which had 2 decimals varied by 2 decimals.
- parameters which had 1 or no decimals varied by 1 decimal.
- parameters which had 3 decimals varied by 3 decimals.

The last parameter, battery, remains the same throughout because its voltage shouldn't vary significantly over a period of several seconds.

This was all done to keep the output as realistic as possible, but at the same time, show second by second changes in the output for the demo.

git-svn-id: svn://localhost/netbeams/trunk@565 559edde2-a4d4-44e0-96fc-4035e40d7396