This revision adds support to mongoDB 1.1.1, which is supposed to add development support to Database Shards in Alpha 2 version. This is not supposed to be running in production environment.
Updates to the Data Persistence is also done for the new Java Driver API, that is
in sync with the new mongoDB version. The updates to the executing experiment script
ran as it was running with the previous version.

Next revisions should add support to Database Sharding just by adding command-lines
calls to the experiment script. In case it runs successfully, the updates will be
committed.

git-svn-id: svn://localhost/netbeams/branches/marcello/persistence@612 559edde2-a4d4-44e0-96fc-4035e40d7396