This log was captured during the experiment described on previous revision. It was captured on the foreign host that maintained the foreign shard for mongoDB.
During the data insertion, mongoDB selected this host to persist
values based on the range of the key latitude, which identifies the
position of sensor. The values were distributed randomically.

Although this foreign mongo shard instance received data, the final
result of data was unbalanced. MongoDB currently used is under
revision Alpha 2, and better support is expeted to be on Alpha 3. I
have sent several emails to the mongoDB list regarding the situation,
so google the mongoDB google groups for details...



git-svn-id: svn://localhost/netbeams/branches/marcello/persistence@615 559edde2-a4d4-44e0-96fc-4035e40d7396