Updates to chapter 2 through 7.
Chapter 2 lost the KVP section, since this approach had not been identified in
the literature for sensor networks. In this way, this makes the primary finding
and contribution of the dissertation. It will only appear during the technology
selection, what is the direct representation of what had happened;

chapter 3 adds details to the taxonomies;

Chapter 4 was cut short to describe SF-BEAMS, NetBEAMS in separate, as well as
keeps the entire implementation to the last chapter (appendix). 

Chapter 5 is the selection of the technology, but still misses highlights regarding
the use of KVP as being a new use.

Chapter 6 and 7 are still the design and implementation chapters. Maybe chapter 7 is
going to be pushed into chapter appendix, but still not sure about it.

Fixing diagrams and adding few others for the remaining chapters.

git-svn-id: svn://localhost/netbeams/branches/marcello/persistence@661 559edde2-a4d4-44e0-96fc-4035e40d7396