Major revision of the thesis structure, and review of chapters, moving chapter 7 to the appendix, and adding the first revision of the abstract in the preamble.
Chap 3. Reviewed the structure, misspelled words, sentences, etc;

Chap 4. Changes to the structure of the case study. The documentation regarding netbeams was pushed to the last chapter: appendix;

Chap 5. The selection of the technology follows the review with Prof. Puder. One section describes each of the technologies briefly, and details their characteristics during the analysis of each of the taxonomies;

Chap 6. Describes the implementation of the data persistence. Still the same, but adding the last chapter the deployment diagram. Implementation details in chapter 7 were pushed to Chapter 10: appendix.

Chap 7. Everything removed and added to the appendix; Changes made to the thesis.tex removed this chapter;

Chap 8. The new Chapter 7, contains the idea of the text, without the experiments results, which will be added after running the experiments with a new revision from mongoDB that is supposed to work better than previous versions;

git-svn-id: svn://localhost/netbeams/branches/marcello/persistence@662 559edde2-a4d4-44e0-96fc-4035e40d7396