Updating the DSP Data Persistence to fix final bugs, and upgrade mongoDB client, since the last JAVA API was not backwards compatible.
DSPDataPersistence, TransientPersistenceLayer: ConcurrentModificationExcetions were the major
cause of this update. Decided to remove the unnecessary hash map, and included a ConcurrentSet
instead. The problem was caused by the following:

1. The transient persistence layer runs concurrently to receive the messages delivered from 
the broker and to send be available for the DSPMongCRUDService to capture the transient 
(in-memory) messages. The insertion of the PersistentMessageUnit into the transient message
was not thread-safe, since the reference for the cache was being returned by the method that
returns the transient messages. By doing so, when DSPMOngoCRUDService tried to acquire the
lock to he collection, the PersistentMessageUnit had already maintained the lock (the main 
cache set). 
   - Return a new set with the copies of what is temporarily in transient cache. This way, the
     lock is not imposed;
   - Any method can use the copies of messages, as they are not locked;
   - All the access to the main collection must be synchronized before executing. The method
     to remove an element from the set now maintains the lock synchronized for the collection.
     
2. Updating the use of Mongo class, using the IP and PORT in separate. For this reason, the
   bootstrap message now adds them in separate;
   
3. Changes to the experiment to support the same changes as well. All tested located.

git-svn-id: svn://localhost/netbeams/branches/marcello/persistence@678 559edde2-a4d4-44e0-96fc-4035e40d7396