Revision of Prof. Puder's audio feedback of chapter 5, as well as entire restructure of the document, such as using more page width and height, making it more visual. Table references not working automatically, as well as page numbers (will review later). Table references are entered manually in the text.
The structure of chapter 5 changed according to the fact that some
sections, subsections were very short, with 1 paragraph. So, I decided
to use two different options:

* If there are more to be discussed, a new advantages/disadvantages
subsection is created; or else, one section includes two paragraphs 
that discusses the adv/disad of the topic;

For the feedback changes, here are the ones requested:

- Adding 2 more sentences to different technologies
- Adding notes on the type of partitioning (KVP uses only Horizontal, since data model is denormalized)
- removing the reference to directory-based approach. Since it is similar to the Key approach. (I have seen examples of the implementation on the application)
- Fixing the sentence about time-series changes on tinyDB
- materialized version = a view of a given state of the database
- trying to decrease the sentence about KVP in relational database models to be better understood.
- Fixing references to the kVP image missing

Finally, I have moved texts from the data model, which basically was
already discussing the decision of the selected technology. I moved it
to the general discussion session. IMO, it makes more sense in this way.

git-svn-id: svn://localhost/netbeams/branches/marcello/persistence@694 559edde2-a4d4-44e0-96fc-4035e40d7396